requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
alembic
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pytest
httpx